Shortest, with BufferedReader
BufferedReader has a method lines(), that returns a Stream<String>,
which can be sent a foreach() method, and we can pass to it a consumer
that fills a TreeSet<String> that is filled with the whole set of lines
from the file.  Once filled, we only need to apply System.out::println
to each String, so the sorted lines are printed with a new line in
between.